Move `ReaLayoutAnimator` to a separate file
